+++ src/Facades/RapidezStatamic.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Rapidez\Statamic\Facades;
+
+use Illuminate\Support\Facades\Facade;
+use Rapidez\Statamic\RapidezStatamic as RapidezStatamicFacade;
+
+/**
+ * @see \Rapidez\Statamic\RapidezStatamic
+ */
+class RapidezStatamic extends Facade
+{
+    protected static function getFacadeAccessor()
+    {
+        return RapidezStatamicFacade::class;
+    }
+}
